Auto-Fuzz: Extract Fuzz Target from generator
